requirements.txt
docling>=1.0.0
PyMuPDF>=1.23.0
python-docx>=0.8.11
layoutparser>=0.3.4
tabula-py>=2.8.0
camelot-py[cv]>=0.10.0
Pillow>=10.0.0
opencv-python>=4.8.0
langchain>=0.1.0
langchain-community>=0.0.10
chromadb>=0.4.0
sentence-transformers>=2.2.2
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
tqdm>=4.65.0
pandas>=2.0.0
numpy>=1.24.0